+++ Fallen-8/Index/DictionaryIndex.cs
@@ -291,7 +291,7 @@ namespace NoSQL.GraphDB.Index
                     var valueCount = reader.ReadInt32();
                     for (var j = 0; j < valueCount; j++)
                     {
-                        var graphElementId = reader.ReadInt32();
+                        var graphElementId = reader.ReadInt64();
                         AGraphElement graphElement;
                         if(fallen8.TryGetGraphElement(out graphElement, graphElementId))
                         {
