+++ Fallen-8/Index/Fulltext/RegExIndex.cs
@@ -474,7 +474,7 @@ namespace NoSQL.GraphDB.Index.Fulltext
                     var valueCount = reader.ReadInt32();
                     for (var j = 0; j < valueCount; j++)
                     {
-                        var graphElementId = reader.ReadInt32();
+                        var graphElementId = reader.ReadInt64();
                         AGraphElement graphElement;
                         if (fallen8.TryGetGraphElement(out graphElement, graphElementId))
                         {
